+++ projects/hipsparse/clients/include/hipsparse_test_unique_ptr.hpp
@@ -342,6 +342,24 @@ namespace hipsparse_test
     };
 #endif
 
+#if(!defined(CUDART_VERSION) || CUDART_VERSION >= 13030)
+    struct spgeam_struct
+    {
+        hipsparseSpGEAMDescr_t descr;
+        spgeam_struct()
+        {
+            hipsparseStatus_t status = hipsparseSpGEAM_createDescr(&descr);
+            verify_hipsparse_status_success(status, "ERROR: spgeam_struct constructor");
+        }
+
+        ~spgeam_struct()
+        {
+            hipsparseStatus_t status = hipsparseSpGEAM_destroyDescr(descr);
+            verify_hipsparse_status_success(status, "ERROR: spgeam_struct destructor");
+        }
+    };
+#endif
+
 } // namespace hipsparse_test
 
 using hipsparse_unique_ptr = std::unique_ptr<void, void (*)(void*)>;